chore: rename the `DigitalTime` component to `DigitalClock`
