+++ edge-apps/clock/src/components/DigitalClock.vue
