+++ edge-apps/clock/src/components/index.ts
@@ -1,3 +1,3 @@
-export { default as DigitalTime } from '@/components/DigitalTime.vue'
+export { default as DigitalClock } from '@/components/DigitalClock.vue'
 export { default as DateDisplay } from '@/components/DateDisplay.vue'
 export { default as BrandLogoCard } from '@/components/BrandLogoCard.vue'
